requirements.txt
chromadb==0.5.5
sentence-transformers==3.0.1
transformers>=4.41.0
pdfplumber==0.11.4
numpy>=1.23
tqdm>=4.66.0
pandas>=2.1.0
google-generativeai>=0.8.5
